# backend/src/design/api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Learnify API
  version: "3.0"
  description: |
    Online Course Platform API — v3.
    Changes from v2:
    - Removed dead endpoints: /courses/featured, /courses/continue, /courses/stats, /lessons/:id/course, /payments/webhook-logs
    - Password validation: min 8 chars, 1 uppercase, 1 number
    - Email uniqueness check on profile update (409 if taken)
    - Pagination params clamped (page >= 1, limit 1-100)
    - Notifications response now includes pagination object
    - Course list: added enrollment, price, instructor_id filters
    - Webhook: improved auth validation, atomic payment update
    - Comments: lesson existence check, batched mention queries, content type validation
    - Bookmarks: course existence check (404)
    - Enrollments: progress object validation
    - Course delete: cascade deletes comments, mentions, notes, reviews, bookmarks
    - Upload: file type validation (MIME whitelist)
servers:
  - url: https://asm3-cos30043.vercel.app
    description: Production (Vercel)
  - url: http://localhost:3000
    description: Local dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id: { type: string, format: uuid }
        instructor_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        difficulty: { type: string, enum: [beginner, intermediate, advanced] }
        price: { type: number }
        thumbnail_url: { type: string, nullable: true }
        instructor_name: { type: string }
        like_count: { type: integer }
        lesson_count: { type: integer }
        enrollment_count: { type: integer }
        avg_rating: { type: number }
        review_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Lesson:
      type: object
      properties:
        id: { type: string, format: uuid }
        course_id: { type: string, format: uuid }
        title: { type: string }
        video_url: { type: string }
        description: { type: string, nullable: true }
        order_index: { type: integer }
    Enrollment:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        course_id: { type: string, format: uuid }
        progress: { type: object }
        purchased_at: { type: string, format: date-time }
        title: { type: string }
        category: { type: string }
        instructor_name: { type: string }
...
    Pagination:
      type: object
      description: "page >= 1, limit clamped to 1-100"
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer }
        pages: { type: integer }
...
paths:
...
  /api/courses:
    get:
      tags: [Courses]
      summary: List courses with search, filter, sort, pagination
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Search title, description, category
        - name: category
          in: query
          schema: { type: string }
        - name: difficulty
          in: query
          schema: { type: string, enum: [beginner, intermediate, advanced] }
        - name: price
          in: query
          schema: { type: string, enum: [free, paid] }
          description: Filter by pricing
        - name: enrollment
          in: query
          schema: { type: string, enum: [enrolled, not-enrolled] }
          description: Filter by enrollment status (requires auth)
        - name: instructor_id
          in: query
          schema: { type: string, format: uuid }
          description: Filter by instructor
        - name: sort
          in: query
          schema: { type: string, enum: [newest, popular, rating, price-asc, price-desc], default: newest }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Course list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      courses:
                        type: array
                        items: { $ref: "#/components/schemas/Course" }
                      pagination: { $ref: "#/components/schemas/Pagination" }

    post:
      tags: [Courses]
      summary: Create course (instructor only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, category, difficulty]
              properties:
                title: { type: string }
                description: { type: string }
                category: { type: string }
                difficulty: { type: string, enum: [beginner, intermediate, advanced] }
                price: { type: number, default: 0 }
                thumbnail_url: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/Course" }
        "403": { description: Not instructor }
  /api/courses/mine:
    get:
      tags: [Courses]
      summary: Instructor's own courses with stats
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Instructor courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      courses:
                        type: array
                        items:
                          allOf:
                            - $ref: "#/components/schemas/Course"
                            - type: object
                              properties:
                                student_count: { type: integer }
        "403": { description: Not instructor }
  /api/courses/{id}:
    get:
      tags: [Courses]
      summary: Course detail with lessons, likes, enrollment, bookmark status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Course"
                      - type: object
                        properties:
                          lessons:
                            type: array
                            items: { $ref: "#/components/schemas/Lesson" }
                          instructor_bio: { type: string, nullable: true }
                          instructor_avatar: { type: string, nullable: true }
                          user_liked: { type: boolean }
                          is_enrolled: { type: boolean }
                          user_bookmarked: { type: boolean }
        "404": { description: Not found }

    put:
      tags: [Courses]
      summary: Update course (owner only)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                category: { type: string }
                difficulty: { type: string }
                price: { type: number }
                thumbnail_url: { type: string }
      responses:
        "200": { description: Updated }
        "403": { description: Not owner }
        "404": { description: Not found }

    delete:
      tags: [Courses]
      summary: Delete course (owner only)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Deleted }
        "403": { description: Not owner }
        "404": { description: Not found }
  /api/courses/{id}/like:
    post:
      tags: [Courses]
      summary: Toggle like
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      liked: { type: boolean }
                      like_count: { type: integer }

  # ── Lessons ──────────────────────────────────────────────
...
  /api/enrollments:
    get:
      tags: [Enrollments]
      summary: My enrolled courses
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Enrollment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Enrollment" }

    post:
      tags: [Enrollments]
      summary: Enroll in course
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [course_id]
              properties:
                course_id: { type: string, format: uuid }
                payment_id: { type: string, format: uuid }
      responses:
        "201": { description: Enrolled }
        "400": { description: Payment not verified }
        "409": { description: Already enrolled }
  /api/enrollments/{id}/progress:
    put:
      tags: [Enrollments]
      summary: Update watch progress
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                progress: { type: object }
      responses:
        "200": { description: Updated }
        "403": { description: Not owner }

  # ── Payments ─────────────────────────────────────────────
...